Refactor: Split components in Management
