+++ src/routes/Management/AdCards/adCards.module.scss
@@ -1,0 +1,26 @@
+@use 'styles/constants/colors';
+
+.cards {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 20px;
+  margin-top: 40px;
+}
+
+@media screen and (min-width: 900px) {
+  .cards {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media screen and (min-width: 1280px) {
+  .cards {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+@media screen and (min-width: 1680px) {
+  .cards {
+    grid-template-columns: repeat(4, 1fr);
+  }
+}
