--- src/routes/Management/editableBox.module.scss
@@ -1,20 +1,0 @@
-@use 'styles/constants/colors';
-@use 'styles/mixins/typo';
-
-.container {
-  display: flex;
-  align-items: center;
-
-  input {
-    @include typo.body14;
-    display: block;
-    padding: 2px 4px;
-    border-bottom: 1px solid colors.$SECONDARY_02;
-  }
-}
-
-.editBox {
-  display: flex;
-  gap: 4px;
-  align-items: center;
-}
